+++ src/components/Routes.tsx
@@ -3,18 +3,18 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
 
 import SuspenseFallback from './SuspenseFallback'
 
-const Home = lazy(() => import('./components/home/Home'))
+const Home = lazy(() => import('./home/Home'))
 
-const BanList = lazy(() => import('./components/banlist/BanList'))
+const BanList = lazy(() => import('./banlist/BanList'))
 
-const ProductBrowse = lazy(() => import('./components/product/ProductBrowse'))
-const ProductInfo = lazy(() => import('./components/product/ProductInfo'))
+const ProductBrowse = lazy(() => import('./product/ProductBrowse'))
+const ProductInfo = lazy(() => import('./product/ProductInfo'))
 
-const Browse = lazy(() => import('./components/card/CardBrowse'))
-const Card = lazy(() => import('./components/card/card-information/CardInformation'))
+const Browse = lazy(() => import('./card/CardBrowse'))
+const Card = lazy(() => import('./card/card-information/CardInformation'))
 
-const About = lazy(() => import('./components/about/About'))
-const HttpErr = lazy(() => import('./components/util/exception/HttpErr'))
+const About = lazy(() => import('./about/About'))
+const HttpErr = lazy(() => import('./util/exception/HttpErr'))
 
 const NAME_maps_ROUTE: { [key: string]: string } = {
 	Home: '/',
