+++ src/components/header-footer/SubNav.tsx
@@ -1,63 +1,39 @@
 import { Typography, Link } from '@mui/material'
 
-import { NAME_maps_ROUTE } from '../../Routes'
+import { NAME_maps_ROUTE } from '../Routes'
 
 export default function SubNav() {
-	return(
-		<div
-			className='scrollable-nav' >
-			<Link
-				underline='none'
-				color='inherit'
-				href={ NAME_maps_ROUTE.Home } >
-				<Typography
-					className='nav-button'
-					color='inherit'
-					>
+	return (
+		<div className='scrollable-nav'>
+			<Link underline='none' color='inherit' href={NAME_maps_ROUTE.Home}>
+				<Typography className='nav-button' color='inherit'>
 					Home
 				</Typography>
 			</Link>
 
-			<Link
-				underline='none'
-				color='inherit'
-				href={ NAME_maps_ROUTE.BanList } >
-				<Typography
-					className='nav-button' color='inherit' >
+			<Link underline='none' color='inherit' href={NAME_maps_ROUTE.BanList}>
+				<Typography className='nav-button' color='inherit'>
 					Ban List
 				</Typography>
 			</Link>
 
-			<Link
-				underline='none'
-				color='inherit'
-				href={ NAME_maps_ROUTE.CardBrowse } >
-				<Typography
-					className='nav-button' color='inherit' >
+			<Link underline='none' color='inherit' href={NAME_maps_ROUTE.CardBrowse}>
+				<Typography className='nav-button' color='inherit'>
 					Card Browse
 				</Typography>
 			</Link>
 
-			<Link
-				underline='none'
-				color='inherit'
-				href={ NAME_maps_ROUTE.ProductBrowse } >
-				<Typography
-					className='nav-button' color='inherit' >
+			<Link underline='none' color='inherit' href={NAME_maps_ROUTE.ProductBrowse}>
+				<Typography className='nav-button' color='inherit'>
 					Product Browse
 				</Typography>
 			</Link>
 
-			<Link
-				underline='none'
-				color='inherit'
-				href={ NAME_maps_ROUTE.About } >
-				<Typography
-					className='nav-button' color='inherit' >
+			<Link underline='none' color='inherit' href={NAME_maps_ROUTE.About}>
+				<Typography className='nav-button' color='inherit'>
 					About
 				</Typography>
 			</Link>
-
 		</div>
 	)
 }
