+++ src/index.tsx
@@ -3,8 +3,8 @@ import { ThemeProvider } from '@mui/material/styles'
 
 import Footer from './components/header-footer/Footer'
 import NavigationBar from './components/header-footer/NavigationBar'
-import Routes from './Routes'
-import MuiTheme from './MuiTheme'
+import Routes from './components/Routes'
+import MuiTheme from './components/MuiTheme'
 
 render(
 	<ThemeProvider theme={MuiTheme}>
